@@ -123,12 +134,15 @@
     "firebase": "^11.6.0",
     "framer-motion": "^12.6.5",
     "graphql-tag": "^2.12.6",
-    "i18next": "^23.10.1",
+    "i18next": "^23.16.8",
+    "i18next-browser-languagedetector": "^8.0.4",
+    "i18next-http-backend": "^3.0.2",
     "idb": "^8.0.2",
     "ioredis": "^5.6.1",
     "lucide-react": "^0.487.0",
-    "next": "15.2.5",
+    "next": "^15.3.0",
     "next-auth": "^4.24.11",
+    "next-csrf": "^0.2.1",
     "next-i18next": "^15.2.0",
     "next-pwa": "^5.6.0",
     "node-cron": "^3.0.3",
