@@ -19,7 +19,6 @@ import { AlertService } from '@/services/alert-service';
 import { ProductService } from '@/services/product-service';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { AlertCircle, Bell, Check, Edit, Plus, Trash2 } from 'lucide-react';
-import { Skeleton } from '@/components/ui/skeleton';
 
 // Alert form schema
 const alertFormSchema = z.object({
@@ -261,10 +260,9 @@ export function AlertsDashboard() {
                 <div className="space-y-2">
                   {[...Array(3)].map((_, i) => (
                     <div key={i} className="flex items-center space-x-4">
-                      <Skeleton className="h-12 w-12 rounded-full" />
                       <div className="space-y-2">
-                        <Skeleton className="h-4 w-[250px]" />
-                        <Skeleton className="h-4 w-[200px]" />
+                        <div className="h-4 w-[250px] bg-gray-200 rounded-full"></div>
+                        <div className="h-4 w-[200px] bg-gray-200 rounded-full"></div>
                       </div>
                     </div>
                   ))}
